[refactor] Move relay into ndk
Now that we are closer to release, let's clean up file structure
