+++ .gitignore
@@ -13,3 +13,4 @@ test-results.xml
 relay/deploy/overlays/staging/secrets.yaml
 .vscode/
 scripts/check-merged*
+.tmp
