+++ conftest.py
@@ -26,15 +26,15 @@ import ssl
 import pytest
 import websockets
 
-from ndk.repos.event_repo import protocol_handler, relay_event_repo
-from relay import (
+from ndk.relay import (
     event_handler,
     message_dispatcher,
     message_handler,
-    server,
     subscription_handler,
 )
-from relay.event_repo import memory_event_repo
+from ndk.relay.event_repo import memory_event_repo
+from ndk.repos.event_repo import protocol_handler, relay_event_repo
+from relay import server
 
 
 def pytest_addoption(parser):
