+++ ndk/relay/__init__.py
@@ -18,23 +18,3 @@
 # WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
 # FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 # OTHER DEALINGS IN THE SOFTWARE.
-
-from ndk.event import event, event_filter
-from relay.event_repo import event_repo
-
-
-class EventHandler:
-    _repo: event_repo.EventRepo
-
-    def __init__(self, repo: event_repo.EventRepo):
-        self._repo = repo
-
-    async def handle_event(self, ev: event.SignedEvent):
-        await self._repo.add(ev)
-
-        if ev.kind in [event.EventKind.SET_METADATA, event.EventKind.CONTACT_LIST]:
-            existing_evs = await self._repo.get(
-                [event_filter.EventFilter(authors=[ev.pubkey], kinds=[ev.kind])]
-            )
-            for e in existing_evs[1:]:  # first entry (latest) is the one we just added
-                await self._repo.remove(e.id)
