+++ ndk/relay/test_message_dispatcher.py
@@ -26,13 +26,13 @@ import pytest
 
 from ndk import serialize
 from ndk.messages import close, eose, event_message, message_factory, notice, request
-from relay import (
+from ndk.relay import (
     event_handler,
     message_dispatcher,
     message_handler,
     subscription_handler,
 )
-from relay.event_repo import memory_event_repo
+from ndk.relay.event_repo import memory_event_repo
 
 
 @pytest.fixture
